Moved away from macroquad frame timing
Implemented frame timing using `SystemTime` and `Duration` rather than
macroquad's `get_frame_time` function. This removes the need to call
`next_frame` every time the main loop iterates in order to properly time
frames. This makes it possible to only render when the next generation
is calculated.

+++ src/args.rs
@@ -17,9 +17,9 @@ pub struct AutomataArgs {
     #[arg(
         short = 'g',
         long = "gens-per-sec",
-        default_value_t = 10.0,
+        default_value_t = 10,
     )]
-    pub gens_per_sec: f32,
+    pub gens_per_sec: u32,
 
     #[command(subcommand)]
     pub cell: CellType,
